pyproject.toml
[project]
name = "equity-research-agent"
version = "0.1.0"
description = "Real-Time Equity Research Agent - AI-powered financial analysis with LangGraph"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Nolan C.", email = "contact@example.com" }]
keywords = ["finance", "ai", "langchain", "langgraph", "rag", "equity-research"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core LLM & Agents
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-groq>=0.2.0",  # Free tier LLM
    "langchain-openai>=0.2.0",  # Optional paid
    "langsmith>=0.1.0",
    
    # Financial Data
    "yfinance>=0.2.40",
    "sec-edgar-downloader>=5.0.0",
    
    # RAG & Vector DB
    "qdrant-client>=1.9.0",
    "unstructured[pdf]>=0.15.0",
    # Using Azure OpenAI embeddings instead of local sentence-transformers
    
    # Web Search
    "duckduckgo-search>=6.0.0",
    
    # API & Backend
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    
    # Caching & Rate Limiting
    "redis>=5.0.0",
    "slowapi>=0.1.9",
    
    # Utilities
    "httpx>=0.27.0",
    "tenacity>=8.4.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.0.0",
    
    # Telegram Bot
    "python-telegram-bot>=21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]

[project.urls]
Homepage = "https://github.com/nolancacheux/equity-research-agent"
Repository = "https://github.com/nolancacheux/equity-research-agent"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "class.*\\(Protocol\\):",
]
